feat: idk

+++ models/wishListsModel.go
@@ -8,7 +8,7 @@ import (
 type WishList struct {
 	gorm.Model
 	IdUser    uint
-	Total     float64
+	Total     int
 	CreatedAt time.Time `gorm:"type:datetime;not null"`
 	UpdatedAt time.Time `gorm:"type:datetime;not null"`
 }
